Rename Logs.js to LogsRoute.js
